Various small fixes
- Reduce size of some enums
- Fix formatting in Editor
- Remove Screenshot, RenderSelectionNextFrame from SceneRenderer
- Add constructor with AddressingMode param in Texture and Texture2D when specifying an existing data source.

[no ci]

+++ Saturn/src/Saturn/ImGui/EntitySelectionReason.h
@@ -30,7 +30,7 @@
 
 namespace Saturn {
 
-	enum EntitySelectionReason : uint32_t
+	enum EntitySelectionReason : uint8_t
 	{
 		ESR_SceneHierarchyPanel = 0x01,
 		ESR_Viewport = 0x02,
